requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
discord==2.2.3
discord.py==2.2.3
emoji==2.2.0
executing==1.2.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
idna==3.4
ipdb==0.13.13
ipython==8.13.2
jedi==0.18.2
jellyfish==0.11.2
Levenshtein==0.21.0
matplotlib-inline==0.1.6
multidict==6.0.4
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
psycopg2==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
python-Levenshtein==0.21.0
rapidfuzz==3.0.0
requests==2.30.0
six==1.16.0
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
urllib3==2.0.2
wcwidth==0.2.6
yarl==1.9.2
